Prototype user id'ing, use github pages for update check now
Still need to test is segments will export okay while using this.
I think they should be fine, but I am adding extra shit onto the
mesh file and I want to make sure it still works after this. Also
I've not tested that the protection works as planned but it's going
to be fucking great!